feat(processors): unify analytical functions and reorganize modules
- Refactored and unified analytical functions in `src/analytical.py`
- Removed deprecated `src/analytical_v2.py`
- Moved and renamed processor modules into `src/processors/` and `src/utils/`
- Added new processor: `src/processors/docling_analytical.py`
- Updated imports and structure in `src/main.py`
- Added `pytest.ini` for test configuration
- Moved and renamed test files to match new module structure
- Updated `pyproject.toml` for new module paths
