+++ pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+pythonpath = src
+asyncio_default_fixture_loop_scope = session
+env_files =
+  .env
+  PYTHONPATH=src/
